instagram/requirements.txt
django~=4.1.4
pillow
django-debug-toolbar
django-extensions
django-bootstrap4